Move wallet_is_synced
