+++ tests/core/full_node/test_full_node.py
@@ -43,7 +43,7 @@ from tests.blockchain.blockchain_test_utils import (
     _validate_and_add_block,
     _validate_and_add_block_no_error,
 )
-from tests.conftest import wallet_is_synced
+from tests.util.wallet_is_synced import wallet_is_synced
 from tests.connection_utils import add_dummy_connection, connect_and_get_peer
 from tests.core.full_node.stores.test_coin_store import get_future_reward_coins
 from tests.core.full_node.test_mempool_performance import wallet_height_at_least
