+++ tests/wallet/test_wallet.py
@@ -22,7 +22,7 @@ from chia.wallet.wallet_node import WalletNode
 from chia.wallet.wallet_state_manager import WalletStateManager
 from chia.simulator.block_tools import BlockTools
 from chia.simulator.time_out_assert import time_out_assert, time_out_assert_not_none
-from tests.conftest import wallet_is_synced
+from tests.util.wallet_is_synced import wallet_is_synced
 from tests.wallet.cat_wallet.test_cat_wallet import tx_in_pool
